removing Video Mode Toggle code.
