+++ Source/OEXAnalytics+Swift.swift
@@ -64,25 +64,4 @@ extension OEXAnalytics {
         event.label = courseId
         return event
     }
-
-    func trackOutlineModeChanged(mode : CourseOutlineMode) {
-        let event = OEXAnalyticsEvent()
-        event.name = OEXAnalyticsEventOutlineModeChanged
-        event.displayName = "Switch outline mode"
-        event.category = OEXAnalyticsCategoryNavigation
-        
-        let modeValue : String
-        
-        switch mode {
-        case .Full:
-            modeValue = OEXAnalyticsValueNavigationModeFull
-            event.label = "Switch to Full Mode"
-        case .Video:
-            modeValue = OEXAnalyticsValueNavigationModeVideo
-            event.label = "Switch to Video Mode"
-        }
-        let info = [OEXAnalyticsKeyNavigationMode : modeValue]
-        self.trackEvent(event, forComponent: nil, withInfo: info)
-    }
-    
 }
